+++ common/src/main/java/io/github/reserveword/imblocker/common/gui/Rectangle.java
@@ -3,7 +3,7 @@ package io.github.reserveword.imblocker.common.gui;
 public record Rectangle(int x, int y, int width, int height) {
 	public static final Rectangle EMPTY = new Rectangle(0, 0, 0, 0);
 	
-	public Rectangle(double scaleFactor, int x, int y, int width, int height) {
+	public Rectangle(double scaleFactor, double x, double y, double width, double height) {
 		this((int) (scaleFactor * x), (int) (scaleFactor * y), 
 				(int) (scaleFactor * width), (int) (scaleFactor * height));
 	}
